# projetos/estoque/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Projeto API do Estoque",
    "description": "Projeto para aplicar os conhecimentos da disciplina de POO2.",
    "version": "0.0.1-SNAPSHOT",
    "contact": {
      "name": "José A. Q. C. Gomes",
      "email": "jose.comparotto@gmail.com",
      "url": "https://github.com/JoseComparotto/poo2"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080/api/estoque"
    }
  ],
  "tags": [
    {
      "name": "ClasseProduto",
      "description": "Macro-agrupador de produtos"
    }
  ],
  "paths": {
    "/classeproduto": {
      "get": {
        "description": "Retorna uma lista com todas as classes de produtos.",
        "operationId": "GetAllClasseProduto",
        "tags": [
          "ClasseProduto"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClasseProdutoArray"
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Cadastra uma nova classe de produto.",
        "tags": [
          "ClasseProduto"
        ],
        "operationId": "CreateClasseProduto",
        "requestBody":{
          "required": true,
          "content": {
            "application/json":{
              "schema":{
                "$ref": "#/components/schemas/ClasseProdutoRequest"
              }
            } 
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json":{
                "schema":{
                  "$ref": "#/components/schemas/ClasseProduto"
                }
              }
            }
          }
        }
      }
    },
    "/classeproduto/{id}": {
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }
      ],
      "get": {
        "description": "Retorna as informações da classe de produto com o Id especificado.",
        "operationId": "GetClasseProdutoById",
        "tags": [
          "ClasseProduto"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClasseProdutoArray"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Atualiza as informações da classe de produto com o Id especificado.",
        "operationId": "UpdateClasseProdutoById",
        "tags": [
          "ClasseProduto"
        ],
        "requestBody":{
          "required": true,
          "content": {
            "application/json":{
              "schema":{
                "$ref": "#/components/schemas/ClasseProdutoRequest"
              }
            } 
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json":{
                "schema":{
                  "$ref": "#/components/schemas/ClasseProduto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Exclui a classe de produto com o Id especificado.",
        "tags": [
          "ClasseProduto"
        ],
        "operationId": "DeleteClasseProdutoById",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClasseProdutoArray"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ClasseProdutoRequest": {
        "type": "object",
        "properties": {
          "descricao": {
            "type": "string",
            "description": "Descrição da classe"
          }
        }
      },
      "ClasseProduto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único"
          },
          "descricao": {
            "type": "string",
            "description": "Descrição da classe"
          },
          "dataInclusao": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "ClasseProdutoArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ClasseProduto"
        }
      }
    }
  }
}